# backend-spring/HealthCARE-App-mrx-part4/swagger-DOCUMENTATION.json
{
    "openapi":  "3.1.0",
    "info":  {
                 "title":  "HealthCare+ API",
                 "description":  "Documentation officielle de l\u0027API HealthCare+ : authentification JWT, gestion des patients, médecins, rendez-vous et dossiers médicaux.",
                 "contact":  {
                                 "name":  "HealthCare+",
                                 "email":  "support@healthcare.local"
                             },
                 "version":  "1.0.0"
             },
    "servers":  [
                    {
                        "url":  "http://localhost:8080",
                        "description":  "Generated server url"
                    }
                ],
    "security":  [
                     {
                         "bearerAuth":  []
                     }
                 ],
...
    "paths":  {
...
                  "/medcins/{id}":  {
                                        "get":  {
                                                    "tags":  [
                                                                 "Medcin"
                                                             ],
                                                    "operationId":  "get_2",
                                                    "parameters":  [
                                                                       {
                                                                           "name":  "id",
                                                                           "in":  "path",
                                                                           "required":  true,
                                                                           "schema":  {
                                                                                          "type":  "integer",
                                                                                          "format":  "int32"
                                                                                      }
                                                                       }
                                                                   ],
                                                    "responses":  {
                                                                      "200":  {
                                                                                  "description":  "OK",
                                                                                  "content":  {
                                                                                                  "*/*":  {
                                                                                                              "schema":  {
                                                                                                                             "$ref":  "#/components/schemas/MedcinGetDTO"
                                                                                                                         }
                                                                                                          }
                                                                                              }
                                                                              }
                                                                  },
                                                    "security":  [
                                                                     {
                                                                         "bearerAuth":  [

                                                                                        ]
                                                                     }
                                                                 ]
                                                },
                                        "put":  {
                                                    "tags":  [
                                                                 "Medcin"
                                                             ],
                                                    "operationId":  "update_3",
                                                    "parameters":  [
                                                                       {
                                                                           "name":  "id",
                                                                           "in":  "path",
                                                                           "required":  true,
                                                                           "schema":  {
                                                                                          "type":  "integer",
                                                                                          "format":  "int32"
                                                                                      }
                                                                       }
                                                                   ],
                                                    "requestBody":  {
                                                                        "content":  {
                                                                                        "application/json":  {
                                                                                                                 "schema":  {
                                                                                                                                "$ref":  "#/components/schemas/MedcinUpdateDTO"
                                                                                                                            }
                                                                                                             }
                                                                                    },
                                                                        "required":  true
                                                                    },
                                                    "responses":  {
                                                                      "200":  {
                                                                                  "description":  "OK",
                                                                                  "content":  {
                                                                                                  "*/*":  {
                                                                                                              "schema":  {
                                                                                                                             "$ref":  "#/components/schemas/MedcinGetDTO"
                                                                                                                         }
                                                                                                          }
                                                                                              }
                                                                              }
                                                                  },
                                                    "security":  [
                                                                     {
                                                                         "bearerAuth":  [

                                                                                        ]
                                                                     }
                                                                 ]
                                                },
                                        "delete":  {
                                                       "tags":  [
                                                                    "Medcin"
                                                                ],
                                                       "operationId":  "delete_3",
                                                       "parameters":  [
                                                                          {
                                                                              "name":  "id",
                                                                              "in":  "path",
                                                                              "required":  true,
                                                                              "schema":  {
                                                                                             "type":  "integer",
                                                                                             "format":  "int32"
                                                                                         }
                                                                          }
                                                                      ],
                                                       "responses":  {
                                                                         "200":  {
                                                                                     "description":  "OK"
                                                                                 }
                                                                     },
                                                       "security":  [
                                                                        {
                                                                            "bearerAuth":  [

                                                                                           ]
                                                                        }
                                                                    ]
                                                   }
                                    },
...
                  "/patients":  {
                                    "get":  {
                                                "tags":  [
                                                             "Patient"
                                                         ],
                                                "operationId":  "list_2",
                                                "parameters":  [
                                                                   {
                                                                       "name":  "page",
                                                                       "in":  "query",
                                                                       "required":  false,
                                                                       "schema":  {
                                                                                      "type":  "integer",
                                                                                      "format":  "int32",
                                                                                      "default":  0
                                                                                  }
                                                                   },
                                                                   {
                                                                       "name":  "size",
                                                                       "in":  "query",
                                                                       "required":  false,
                                                                       "schema":  {
                                                                                      "type":  "integer",
                                                                                      "format":  "int32",
                                                                                      "default":  5
                                                                                  }
                                                                   },
                                                                   {
                                                                       "name":  "sort",
                                                                       "in":  "query",
                                                                       "required":  false,
                                                                       "schema":  {
                                                                                      "type":  "string",
                                                                                      "default":  "nom"
                                                                                  }
                                                                   },
                                                                   {
                                                                       "name":  "direction",
                                                                       "in":  "query",
                                                                       "required":  false,
                                                                       "schema":  {
                                                                                      "type":  "string",
                                                                                      "default":  "asc"
                                                                                  }
                                                                   }
                                                               ],
                                                "responses":  {
                                                                  "200":  {
                                                                              "description":  "OK",
                                                                              "content":  {
                                                                                              "*/*":  {
                                                                                                          "schema":  {
                                                                                                                         "$ref":  "#/components/schemas/PagePatientGetDTO"
                                                                                                                     }
                                                                                                      }
                                                                                          }
                                                                          }
                                                              },
                                                "security":  [
                                                                 {
                                                                     "bearerAuth":  [

                                                                                    ]
                                                                 }
                                                             ]
                                            },
                                    "post":  {
                                                 "tags":  [
                                                              "Patient"
                                                          ],
                                                 "operationId":  "ajouter",
                                                 "requestBody":  {
                                                                     "content":  {
                                                                                     "application/json":  {
                                                                                                              "schema":  {
                                                                                                                             "$ref":  "#/components/schemas/PatientPostDTO"
                                                                                                                         }
                                                                                                          }
                                                                                 },
                                                                     "required":  true
                                                                 },
                                                 "responses":  {
                                                                   "200":  {
                                                                               "description":  "OK",
                                                                               "content":  {
                                                                                               "*/*":  {
                                                                                                           "schema":  {
                                                                                                                          "$ref":  "#/components/schemas/PatientGetDTO"
                                                                                                                      }
                                                                                                       }
                                                                                           }
                                                                           }
                                                               },
                                                 "security":  [
                                                                  {
                                                                      "bearerAuth":  [

                                                                                     ]
                                                                  }
                                                              ]
                                             }
                                },
                  "/medcins":  {
                                   "get":  {
                                               "tags":  [
                                                            "Medcin"
                                                        ],
                                               "operationId":  "list_3",
                                               "parameters":  [
                                                                  {
                                                                      "name":  "page",
                                                                      "in":  "query",
                                                                      "required":  false,
                                                                      "schema":  {
                                                                                     "type":  "integer",
                                                                                     "format":  "int32",
                                                                                     "default":  0
                                                                                 }
                                                                  },
                                                                  {
                                                                      "name":  "size",
                                                                      "in":  "query",
                                                                      "required":  false,
                                                                      "schema":  {
                                                                                     "type":  "integer",
                                                                                     "format":  "int32",
                                                                                     "default":  5
                                                                                 }
                                                                  },
                                                                  {
                                                                      "name":  "sort",
                                                                      "in":  "query",
                                                                      "required":  false,
                                                                      "schema":  {
                                                                                     "type":  "string",
                                                                                     "default":  "specialite"
                                                                                 }
                                                                  },
                                                                  {
                                                                      "name":  "direction",
                                                                      "in":  "query",
                                                                      "required":  false,
                                                                      "schema":  {
                                                                                     "type":  "string",
                                                                                     "default":  "asc"
                                                                                 }
                                                                  }
                                                              ],
                                               "responses":  {
                                                                 "200":  {
                                                                             "description":  "OK",
                                                                             "content":  {
                                                                                             "*/*":  {
                                                                                                         "schema":  {
                                                                                                                        "$ref":  "#/components/schemas/PageMedcinGetDTO"
                                                                                                                    }
                                                                                                     }
                                                                                         }
                                                                         }
                                                             },
                                               "security":  [
                                                                {
                                                                    "bearerAuth":  [

                                                                                   ]
                                                                }
                                                            ]
                                           },
                                   "post":  {
                                                "tags":  [
                                                             "Medcin"
                                                         ],
                                                "operationId":  "ajouter_1",
                                                "requestBody":  {
                                                                    "content":  {
                                                                                    "application/json":  {
                                                                                                             "schema":  {
                                                                                                                            "$ref":  "#/components/schemas/MedcinPostDTO"
                                                                                                                        }
                                                                                                         }
                                                                                },
                                                                    "required":  true
                                                                },
                                                "responses":  {
                                                                  "200":  {
                                                                              "description":  "OK",
                                                                              "content":  {
                                                                                              "*/*":  {
                                                                                                          "schema":  {
                                                                                                                         "$ref":  "#/components/schemas/MedcinGetDTO"
                                                                                                                     }
                                                                                                      }
                                                                                          }
                                                                          }
                                                              },
                                                "security":  [
                                                                 {
                                                                     "bearerAuth":  [

                                                                                    ]
                                                                 }
                                                             ]
                                            }
                               },
...
                  "/patients/me":  {
                                       "get":  {
                                                   "tags":  [
                                                                "Patient"
                                                            ],
                                                   "operationId":  "getMyProfile",
                                                   "responses":  {
                                                                     "200":  {
                                                                                 "description":  "OK",
                                                                                 "content":  {
                                                                                                 "*/*":  {
                                                                                                             "schema":  {
                                                                                                                            "$ref":  "#/components/schemas/PatientGetDTO"
                                                                                                                        }
                                                                                                         }
                                                                                             }
                                                                             }
                                                                 },
                                                   "security":  [
                                                                    {
                                                                        "bearerAuth":  [

                                                                                       ]
                                                                    }
                                                                ]
                                               }
                                   },
...
              },
    "components":  {
                       "schemas":  {
...
                                       "MedcinGetDTO":  {
                                                            "type":  "object",
                                                            "properties":  {
                                                                               "id":  {
                                                                                          "type":  "integer",
                                                                                          "format":  "int32"
                                                                                      },
                                                                               "nom":  {
                                                                                           "type":  "string"
                                                                                       },
                                                                               "specialite":  {
                                                                                                  "type":  "string"
                                                                                              },
                                                                               "email":  {
                                                                                             "type":  "string"
                                                                                         },
                                                                               "telephone":  {
                                                                                                 "type":  "string"
                                                                                             }
                                                                           }
                                                        },
...
                                   },
                       "securitySchemes":  {
                                               "bearerAuth":  {
                                                                  "type":  "http",
                                                                  "name":  "bearerAuth",
                                                                  "scheme":  "bearer",
                                                                  "bearerFormat":  "JWT"
                                                              }
                                           }
                   }
}